requirements.txt

aiohttp==3.9.3
coremltools>=6.0.0
aiosignal==1.3.1
numpy
langchain_openai
torch==2.0.1
sounddevice
langchain
langchain_openai
kaleido==0.2.1  
scipy
annotated-types==0.6.0
attrs==23.2.0
boto3
botocore==1.34.69
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
coremltools==7.1
dataclasses-json==0.6.4
distlib==0.3.8
einops==0.7.0
encodec==0.1.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.3.1
funcy==2.0
huggingface-hub==0.22.0
identify==2.5.35
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain
langchain-community
langchain-core
langchain-text-splitters==0.0.1
langsmith==0.1.36
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.8.1
nodeenv==1.8.0
numba==0.59.1
numpy==1.26.4
openai-whisper
orjson==3.10.0
packaging==23.2
platformdirs==4.2.0
pre-commit==3.6.2
pyaml==23.12.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
safetensors
scipy==1.12.0
six==1.16.0
sounddevice
SpeechRecognition
SQLAlchemy==2.0.29
suno-bark
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
torch
torchaudio
tqdm==4.66.2
transformers==4.39.1
urllib3
virtualenv
yarl
